Scheduler Almost done
Both MLFQ and RR algorithms are fully written but not tested, yet. The function admit schedule is written as well where its main purpose is to take a client request, create an RCB for that particular request, and then admit it to the scheduler chosen at run time. Remaining is the function that serves request from the scheduler and resubmit them to scheduler if necessary